Unstable tile: Implement a "slow fall" phase for sprites without "shake" and "dissolve" actions.
Sprites with neither a "shake" nor a "dissolve" action now start to fall down
immediately. For the first 0.5 seconds, the gravity modifier is set to 10%
though, so that the movement is *very* slow at first. It is noticable however
and hints at the instability of the tile. Hopefully, players won't be surprised
by a falling platform anymore now.

SVN-Revision: 6612