levels/world2: Add "Bye Bye Forest" (future Forest Island outro)
"Bye Bye Forest" is a new level designed to be placed at the end of
Forest Island, where it will be used as last Forest level and outro. It
features elements of waters inspired by those in the old
"Countercurrent" level. Furthermore, it features a slightly non-linear
path as well as some more modern features not previously used on Forest
Island (e.g. moving coins).
The water signifies how close the coast is, and at the end of the level,
Tux leaves Forest Island using a wooden block moving platform on the
ocean.

Ref: SuperTux/data#69 ("levels/world2: Add "Bye Bye Forest" + "This
Forest Is Haunted"")
[max@teufelsnetz.com: taken from 31a2a457b7924f66af3794f71a7af66d175cbba5,
split off "Bye Bye Forest", and extended commit message]
Reviewed-by: Max Teufel <max@teufelsnetz.com>
Tested-by: Max Teufel <max@teufelsnetz.com>
Cc: alex jockers <alex.jockers@gmx.de>
Cc: Max Teufel <max@teufelsnetz.com>
Cc: Carsten Wirtz <carstenwirtz@live.de>